Split up tst_rast into different modules
It now compiles properly on TOS.

on_curs() and off_curs() were added to tst_rast.c, as they were needed
to prevent a cursor from blinking while staring at images.
