+++ MAKEFILE
@@ -1,6 +1,6 @@
 all: acaddom tests
 
-tests: tst_mdl tst_mse tst_mus tst_psg tst_rast tst_sfx
+tests: tst_bmp tst_mdl tst_mse tst_mus tst_psg tst_shp tst_sfx
 
 tst_mdl: src\model.o src\arg_list.o src\effects.o src\events.o src\input.o src\in_asm.o src\ints_asm.o src\move.o src\psg.o src\rast_asm.o src\super.o src\su_asm.o src\tst_hndl.o src\tst_mdl.o src\move.o src\test.o src\vector.o
 	cc68x -g src\model.o src\arg_list.o src\effects.o src\events.o src\input.o src\in_asm.o src\ints_asm.o src\move.o src\psg.o src\rast_asm.o src\super.o src\su_asm.o src\tst_hndl.o src\tst_mdl.o src\test.o src\vector.o -o bin\tst_mdl.tos
@@ -26,10 +26,19 @@ tst_psg: src\psg.o src\arg_list.o src\input.o src\in_asm.o src\ints_asm.o src\ra
 src\tst_psg.o: src\tst_psg.c src\psg.h src\arg_list.h src\input.h src\raster.h src\super.h src\test.h src\tst_hndl.h src\types.h src\vector.h
 	cc68x -g -c src\tst_psg.c
 
-tst_rast: src\test.o src\tst_hndl.o src\tst_rast.o src\arg_list.o src\bitmaps.o src\input.o src\in_asm.o src\ints_asm.o src\raster.o src\rast_asm.o src\super.o src\su_asm.o src\vector.o
-	cc68x -g src\test.o src\tst_hndl.o src\tst_rast.o src\arg_list.o src\bitmaps.o src\input.o src\in_asm.o src\ints_asm.o src\raster.o src\rast_asm.o src\super.o src\su_asm.o src\vector.o -o bin\tst_rast.tos
+tst_bmp: src\test.o src\tst_hndl.o src\tst_rast.o src\tst_bmp.o src\arg_list.o src\bitmaps.o src\input.o src\in_asm.o src\ints_asm.o src\raster.o src\rast_asm.o src\super.o src\su_asm.o src\vector.o
+	cc68x -g src\test.o src\tst_hndl.o src\tst_rast.o src\tst_bmp.o src\arg_list.o src\bitmaps.o src\input.o src\in_asm.o src\ints_asm.o src\raster.o src\rast_asm.o src\super.o src\su_asm.o src\vector.o -o bin\tst_bmp.tos
 
-src\tst_rast.o: src\tst_rast.c src\bitmaps.h src\raster.h src\scrn.h src\types.h
+src\tst_bmp.o: src\tst_bmp.c src\arg_list.h src\bitmaps.h src\bool.h src\raster.h src\scrn.h src\test.h src\tst_hndl.h src\tst_rast.h src\types.h
+	cc68x -g -c src\tst_bmp.c
+
+tst_shp: src\test.o src\tst_hndl.o src\tst_rast.o src\tst_shp.o src\arg_list.o src\input.o src\in_asm.o src\ints_asm.o src\raster.o src\rast_asm.o src\super.o src\su_asm.o src\vector.o
+	cc68x -g src\test.o src\tst_hndl.o src\tst_rast.o src\tst_shp.o src\arg_list.o src\input.o src\in_asm.o src\ints_asm.o src\raster.o src\rast_asm.o src\super.o src\su_asm.o src\vector.o -o bin\tst_shp.tos
+
+src\tst_shp.o: src\tst_shp.c src\arg_list.h src\bool.h src\raster.h src\scrn.h src\test.h src\tst_hndl.h src\tst_rast.h src\types.h
+	cc68x -g -c src\tst_shp.c
+
+src\tst_rast.o: src\tst_rast.c src\tst_rast.h src\arg_list.h src\bool.h src\input.h src\raster.h src\scrn.h src\types.h
 	cc68x -g -c src\tst_rast.c
 
 tst_sfx: src\tst_sfx.o src\psg.o src\arg_list.o src\effects.o src\input.o src\in_asm.o src\ints_asm.o src\rast_asm.o src\super.o src\su_asm.o src\test.o src\tst_hndl.o src\vector.o
